Refactor `module` module for reduced complexity
